Move SharedSubmissionUtil into the shared-resources module.  Move the test for SharedSubmissionUtil to fedora-builder to avoid a circular dependency on the builder implementation.
