+++ shared-assembler/pom.xml
@@ -113,6 +113,13 @@
             <artifactId>fedora-builder</artifactId>
             <version>${project.parent.version}</version>
             <scope>test</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>org.dataconservancy.pass.deposit</groupId>
+            <artifactId>fedora-builder</artifactId>
+            <version>${project.parent.version}</version>
+            <scope>test</scope>
             <classifier>tests</classifier>
             <type>test-jar</type>
         </dependency>
