+++ shared-resources/pom.xml
@@ -64,6 +64,13 @@
         </dependency>
 
         <dependency>
+            <groupId>org.dataconservancy.pass.deposit</groupId>
+            <artifactId>builder-api</artifactId>
+            <version>${project.parent.version}</version>
+            <scope>test</scope>
+        </dependency>
+
+        <dependency>
             <groupId>com.fasterxml.jackson.core</groupId>
             <artifactId>jackson-core</artifactId>
             <version>${jackson.version}</version>
